Rename delayedmods to lazymodules
