+++ spyder_kernels/utils/iofuncs.py
@@ -32,10 +32,10 @@ import glob
 
 # Local imports
 from spyder_kernels.py3compat import getcwd, pickle, PY2, to_text_string
-from spyder_kernels.utils.delayedmods import FakeObject, PIL
-from spyder_kernels.utils.delayedmods import numpy as np
-from spyder_kernels.utils.delayedmods import pandas as pd
-from spyder_kernels.utils.delayedmods import scipy as sp
+from spyder_kernels.utils.lazymodules import FakeObject, PIL
+from spyder_kernels.utils.lazymodules import numpy as np
+from spyder_kernels.utils.lazymodules import pandas as pd
+from spyder_kernels.utils.lazymodules import scipy as sp
 
 
 class MatlabStruct(dict):
